[SPARK-46266][PS][CONNECT][TESTS] Re-organize `NumOpsTests`
### What changes were proposed in this pull request?
Re-organize `NumOpsTests`, factor out the `astype` tests

### Why are the changes needed?
group the tests by topics

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44182 from zhengruifeng/ps_reorg_as_type.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
